# PERF: RangeIndex.argmin/argmax