```python
In [1]: import pandas as pd

In [2]: ri = pd.RangeIndex(100_000)

In [3]: %timeit ri.argmin()
651 ns ± 8.43 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # PR

In [3]: %timeit ri.argmin()
19.6 µs ± 73 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)   # main
```